Support arbitrary-length token encoding with Long IDs
The encoded length is now set per token subclass (1-12 chars), replacing
the hardcoded 6-character limit. Upgraded domain size and IDs from Int
to Long throughout the cipher, encoder, and token layers.

Entire-Checkpoint: 6fba74ba99f3

+++ src/main/kotlin/FeistelCipher.kt
@@ -4,31 +4,31 @@ import kotlin.math.sqrt
 /** Permutes integers reversibly within `[0, [domainSize])` using [key], so that sequential inputs produce unrelated outputs. */
 class FeistelCipher(
     private val key: Long,
-    private val domainSize: Int,
+    private val domainSize: Long,
 ) {
     private val rounds = 8
-    private val halfModulus = ceil(sqrt(domainSize.toDouble())).toInt()
+    private val halfModulus = ceil(sqrt(domainSize.toDouble())).toLong()
 
     init {
-        require(halfModulus.toLong() * halfModulus >= domainSize) {
+        require(halfModulus * halfModulus >= domainSize) {
             "Domain split failed: $halfModulus * $halfModulus < $domainSize"
         }
     }
 
-    private fun roundFunction(value: Int, round: Int): Int {
-        var h = value.toLong()
+    private fun roundFunction(value: Long, round: Int): Long {
+        var h = value
         h = h * 2654435761L + key
-        h = h * 2246822519L + round.toLong()
+        h = h * 2246822519L + round
         h = h xor (h ushr 16)
         h = h * 2654435761L
         h = h xor (h ushr 13)
         h = h * 3266489917L
         h = h xor (h ushr 16)
-        return h.mod(halfModulus.toLong()).toInt()
+        return h.mod(halfModulus)
     }
 
     /** Maps [input] to a unique, unrelated integer within the same domain. */
-    fun encrypt(input: Int): Int {
+    fun encrypt(input: Long): Long {
         require(input in 0 until domainSize) { "Input $input outside domain [0, $domainSize)" }
         var value = input
         do {
@@ -46,7 +46,7 @@ class FeistelCipher(
     }
 
     /** Recovers the original integer from an encrypted [input]. */
-    fun decrypt(input: Int): Int {
+    fun decrypt(input: Long): Long {
         require(input in 0 until domainSize) { "Input $input outside domain [0, $domainSize)" }
         var value = input
         do {
@@ -54,8 +54,7 @@ class FeistelCipher(
             var right = value % halfModulus
             for (round in (rounds - 1) downTo 0) {
                 val newRight = left
-                val newLeft = (right.toLong() - roundFunction(left, round).toLong())
-                    .mod(halfModulus.toLong()).toInt()
+                val newLeft = (right - roundFunction(left, round)).mod(halfModulus)
                 left = newLeft
                 right = newRight
             }
